Grouped shared utilities for behave tests
