+++ enginecore/features/steps/ups_battery.py
@@ -1,8 +1,7 @@
+import time
+
 # pylint: disable=no-name-in-module,function-redefined,missing-docstring,unused-import
 from behave import given, when, then, step
-from pysnmp import hlapi
-from circuits import Component, handler
-import time
 
 # plyint: enable=no-name-in-module
 
@@ -13,55 +12,7 @@ from enginecore.state.hardware.asset_events import VoltageDecreased, VoltageIncr
 from enginecore.state.event_map import PowerEventMap
 import enginecore.state.state_initializer as state_ini
 
-
-def query_oid_value(oid, host="localhost", port=1024):
-    """Helper function to query snmp interface of a device"""
-    error_indicator, error_status, error_idx, var_binds = next(
-        hlapi.getCmd(
-            hlapi.SnmpEngine(),
-            hlapi.CommunityData("private", mpModel=0),
-            hlapi.UdpTransportTarget((host, port)),
-            hlapi.ContextData(),
-            hlapi.ObjectType(hlapi.ObjectIdentity(oid)),
-            lookupMib=False,
-        )
-    )
-
-    if error_indicator:
-        print(error_indicator)
-    elif error_status:
-        print(
-            "%s at %s"
-            % (
-                error_status.prettyPrint(),
-                error_idx and var_binds[int(error_idx) - 1][0] or "?",
-            )
-        )
-    else:
-        v_bind = var_binds[0]
-        return v_bind[1]
-
-    return None
-
-
-class FakeEngine(Component):
-    def __init__(self, asset):
-        super(FakeEngine, self).__init__()
-        self._q_events = None
-        asset.register(self)
-        self._asset = asset
-
-    def VoltageDecreased_complete(self, evt, e_result):
-        self.stop()
-
-    def VoltageIncreased_complete(self, evt, e_result):
-        self.stop()
-
-    def queue_event(self, event):
-        self._q_events = event
-
-    def started(self, _):
-        self.fire(self._q_events, self._asset)
+from test_helpers import FakeEngine, query_snmp_interface
 
 
 @given("the system model is empty")
@@ -97,7 +48,7 @@ def step_impl(context, key):
 @when('voltage "{volt}" drops below "{low_threshold}" threshold by "{volt_change:d}"')
 def step_impl(context, low_threshold, volt, volt_change):
     low_th_oid = context.ups.state.get_oid_by_name(low_threshold).oid
-    low_th_value = query_oid_value(low_th_oid)
+    low_th_value = query_snmp_interface(low_th_oid)
 
     assert low_th_value > 0
 
@@ -121,7 +72,7 @@ def step_impl(context, volt):
 @when('voltage "{volt}" spikes above "{high_threshold}" threshold by "{volt_change:d}"')
 def step_impl(context, high_threshold, volt, volt_change):
     high_oid = context.ups.state.get_oid_by_name(high_threshold).oid
-    high_value = query_oid_value(high_oid)
+    high_value = query_snmp_interface(high_oid)
 
     assert high_value > 0
 
@@ -145,7 +96,7 @@ def step_impl(context):
 @then('UPS transfer reason is set to "{t_reason}"')
 def step_impl(context, t_reason):
     transfer_reason_oid = context.ups.state.get_oid_by_name("InputLineFailCause").oid
-    varbind_value = query_oid_value(transfer_reason_oid)
+    varbind_value = query_snmp_interface(transfer_reason_oid)
     print("varbind&t", varbind_value, t_reason)
     assert context.ups.state.get_transfer_reason().name == t_reason
     assert context.ups.state.InputLineFailCause(varbind_value).name == t_reason
